requirements.txt
numpy
torch
opencv-python
omegaconf
matplotlib
seaborn
Pillow
torchvision
scikit-image
tqdm
tensorboard